{{ cookiecutter.project_name }}/requirements.txt
arrow==0.15.7
asgiref==3.2.3
binaryornot==0.4.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cookiecutter==1.7.2
coverage==5.2
Django==3.0.8
django-cors-headers==3.4.0
django-debug-toolbar==2.2
django-extensions==3.0.3
django-model-utils==4.0.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
future==0.18.2
idna==2.8
Jinja2==2.11.2
jinja2-time==0.2.0
MarkupSafe==1.1.1
pathlib==1.0.1
poyo==0.5.0
psycopg2==2.8.5
PyJWT==1.7.1
python-dateutil==2.8.1
python-decouple==3.3
python-slugify==4.0.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.0
style==1.1.6
text-unidecode==1.3
update==0.0.1
urllib3==1.25.7
whichcraft==0.6.1
